activity_test.go
package nmeasentencetomqtt_flogo

import (
	"io/ioutil"
	"testing"

	"github.com/TIBCOSoftware/flogo-lib/core/activity"
	"github.com/TIBCOSoftware/flogo-contrib/action/flow/test"
	"fmt"
)

var activityMetadata *activity.Metadata

func getActivityMetadata() *activity.Metadata {

	if activityMetadata == nil {
		jsonMetadataBytes, err := ioutil.ReadFile("activity.json")
		if err != nil{
			panic("No Json Metadata found for activity.json path")
		}

		activityMetadata = activity.NewMetadata(string(jsonMetadataBytes))
	}

	return activityMetadata
}

func TestCreate(t *testing.T) {

	act := NewActivity(getActivityMetadata())

	if act == nil {
		t.Error("Activity Not Created")
		t.Fail()
		return
	}
}

func TestEval(t *testing.T) {

	defer func() {
		if r := recover(); r != nil {
			t.Failed()
			t.Errorf("panic during execution: %v", r)
		}
	}()

	act := NewActivity(getActivityMetadata())
	tc := test.NewTestActivityContext(getActivityMetadata())

	tc.SetInput("address", "localhost:6379")
	tc.SetInput("dbNo", 0)
	tc.SetInput("sentence", "$GPRMC,133515.283,A,3510.55,S,13828.45,E,32.9,63.0,190419,,,*34'")
	act.Eval(tc)

	message := tc.GetOutput("mqtt_message")
	topic := tc.GetOutput("topic")

	fmt.Println(message)
	fmt.Println(topic)

	//check result attr
}
